refactor: move Deck persistence into entity API
