+++ src/entities/deck/index.ts
@@ -1,15 +1,11 @@
-export { generateDeckId } from "./api/firestore";
+export { createDeck, deleteDeck, editDeck, generateDeckId } from "./api/firestore";
 export { CATEGORY, getCategory, isHighlightLanguage } from "./model/rules";
-export { createDeckSchema, deleteDeckSchema, editDeckSchema } from "./model/schema";
 export { useDeck, useDecks } from "./model/hooks";
 export { clearDecks, replaceDecks } from "./model/store";
 export type {
   Category,
   Deck,
-  DeckCreate,
   DeckCreateInput,
   DeckEdit,
   DeckId,
-  DeleteDeckInput,
-  EditDeckInput,
 } from "./model/types";
