--- src/features/deck/create/api/createDeck.ts
@@ -1,8 +1,0 @@
-import { createDeckSchema, type DeckCreateInput } from "@/entities/deck";
-
-import { createDeckDocument } from "./firestore";
-
-export const createDeck = async (uid: string, deck: DeckCreateInput): Promise<void> => {
-  const input = createDeckSchema.parse({ uid, deck });
-  await createDeckDocument(input.deck);
-};
