--- src/features/deck/create/index.ts
@@ -1,1 +1,0 @@
-export { createDeck } from "./api/createDeck";
