--- src/features/deck/edit/api/editDeck.spec.ts
@@ -1,11 +1,0 @@
-import { describe, expect, it, vi } from "vitest";
-
-vi.mock("@/shared/firebase", () => ({ db: {} }));
-
-import { editDeck } from "./editDeck";
-
-describe("editDeck", () => {
-  it("rejects an edit without a confirmed user", async () => {
-    await expect(editDeck("", { id: "deck" })).rejects.toThrow("confirmed user");
-  });
-});
