+++ src/features/deck/edit/model/useEditDeck.ts
@@ -1,7 +1,6 @@
-import type { DeckEdit } from "@/entities/deck";
+import { type DeckEdit, editDeck } from "@/entities/deck";
 
 import { useAuthSession } from "@/entities/auth";
-import { editDeck } from "../api/editDeck";
 
 export const useEditDeck = () => {
   const auth = useAuthSession();
