+++ src/pages/deck-import/ui/DeckImportPage.spec.tsx
@@ -30,7 +30,7 @@ const mocks = vi.hoisted(() => ({
 }));
 
 vi.mock("react-router-dom", () => ({ useNavigate: () => mocks.navigate }));
-vi.mock("@/entities/deck", () => ({ useDecks: () => [] }));
+vi.mock("@/entities/deck", () => ({ createDeck: vi.fn(), useDecks: () => [] }));
 vi.mock("@/entities/card", () => ({ useCards: () => [] }));
 vi.mock("@/features/deck/import", () => ({
   SAMPLE_CSV_TEXT: "sample csv",
