+++ src/pages/deck-import/ui/DeckImportPage.tsx
@@ -2,13 +2,12 @@ import type React from "react";
 import { useNavigate } from "react-router-dom";
 import { useKey } from "react-use";
 
-import { useDecks } from "@/entities/deck";
+import { createDeck, useDecks } from "@/entities/deck";
 import { useCards } from "@/entities/card";
 import { usePreferences } from "@/entities/preferences";
 import { createCard, generateCardId } from "@/features/card/create";
 import { editCard } from "@/features/card/edit";
 import { useCardReadState } from "@/features/card/read";
-import { createDeck } from "@/features/deck/create";
 import { downloadSampleCsv, SAMPLE_CSV_TEXT, useDeckImport } from "@/features/deck/import";
 import { AppLayout } from "@/widgets/app-layout";
