+++ src/pages/deck-list/ui/DeckListPage.tsx
@@ -3,11 +3,10 @@ import { useNavigate } from "react-router-dom";
 import { useKey } from "react-use";
 
 import { filterCardsByDeckId, useCards } from "@/entities/card";
-import { type Deck, type DeckId, useDecks } from "@/entities/deck";
+import { createDeck, type Deck, type DeckId, useDecks } from "@/entities/deck";
 import { createCard, generateCardId } from "@/features/card/create";
 import { editCard } from "@/features/card/edit";
 import { useCardReadState } from "@/features/card/read";
-import { createDeck } from "@/features/deck/create";
 import { useDeleteDeck } from "@/features/deck/delete";
 import { buildDeckListSections } from "@/features/deck/list";
 import { downloadDeckCsv } from "@/features/deck/export";
