+++ test/integration/firestore/subscriptions.spec.ts
@@ -11,9 +11,7 @@ import { deleteApp, getApps } from "firebase/app";
 import { createCard as createCardCommand } from "@/features/card/create";
 import { deleteCard } from "@/features/card/delete/api/deleteCard";
 import { editCard } from "@/features/card/edit";
-import { createDeck as createDeckCommand } from "@/features/deck/create";
-import { deleteDeck } from "@/features/deck/delete/api/deleteDeck";
-import { editDeck } from "@/features/deck/edit/api/editDeck";
+import { createDeck as createDeckCommand, deleteDeck, editDeck } from "@/entities/deck";
 import { startCardSynchronization } from "@/app/providers/remote-read/card";
 import { subscribeDecks } from "@/app/providers/remote-read/deck";
 import { cardStore } from "@/entities/card/model/store";
